+++ src/js/index.js
@@ -1,62 +1,11 @@
-/**
+/**!
  * copyright (c) 2023 nousantx
  */
 
+import property from "./property.js";
 function makeTenoxUI(element) {
   this.element = element;
-  this.styles = {
-    // Mapping type and its Property
-    p: ["padding"],
-    pt: ["paddingTop"],
-    pb: ["paddingBottom"],
-    pr: ["paddingRight"],
-    pl: ["paddingLeft"],
-    m: ["margin"],
-    mt: ["marginTop"],
-    mb: ["marginBottom"],
-    mr: ["marginRight"],
-    ml: ["marginLeft"],
-    fs: ["fontSize"],
-    fw: ["fontWeight"],
-    z: ["zIndex"],
-    t: ["top"],
-    b: ["bottom"],
-    r: ["right"],
-    l: ["left"],
-    br: ["borderRadius"],
-    bw: ["borderWidth"],
-    "bw-left": ["borderLeftWidth"],
-    "bw-right": ["borderRightWidth"],
-    "bw-top": ["borderTopWidth"],
-    "bw-bottom": ["borderBottomWidth"],
-    w: ["width"],
-    "w-mx": ["maxWidth"],
-    "w-mn": ["minWidth"],
-    h: ["height"],
-    "h-mx": ["maxHeight"],
-    "h-mn": ["minHeight"],
-    fx: ["flex"],
-    gap: ["gap"],
-    ti: ["textIndent"],
-    "backdrop-blur": ["backdropFilter"],
-    ph: ["paddingLeft", "paddingRight"],
-    pv: ["paddingTop", "paddingBottom"],
-    mv: ["marginTop", "marginBottom"],
-    mh: ["marginLeft", "marginRight"],
-    "text-space": ["letterSpacing"],
-    "word-space": ["wordSpacing"],
-    "line-height": ["lineHeight"],
-    "radius-tl": ["borderTopLeftRadius"],
-    "radius-tr": ["borderTopRightRadius"],
-    "radius-bl": ["borderBottomLeftRadius"],
-    "radius-br": ["borderBottomRightRadius"],
-    "radius-top": ["borderTopLeftRadius", "borderTopRightRadius"],
-    "radius-bottom": ["borderBottomLeftRadius", "borderBottomRightRadius"],
-    "radius-left": ["borderTopLeftRadius", "borderBottomLeftRadius"],
-    "radius-right": ["borderTopRightRadius", "borderBottomRightRadius"],
-    opa: ["opacity"],
-    // Other mapping here
-  };
+  this.styles = property;
 }
 
 makeTenoxUI.prototype.applyStyle = function (type, value, unit) {
@@ -64,9 +13,32 @@ makeTenoxUI.prototype.applyStyle = function (type, value, unit) {
 
   if (properties) {
     properties.forEach((property) => {
+      // Filter Custom Property
+      if (property === "filter") {
+        const existingFilter = this.element.style[property];
+        this.element.style[property] = existingFilter
+          ? `${existingFilter} ${type}(${value}${unit})`
+          : `${type}(${value}${unit})`;
+      }
       // Make custom property for flex
-      if (property === "flex") {
+      else if (property === "flex") {
         this.element.style[property] = `1 1 ${value}${unit}`;
+      } // Grid System Property
+      else if (
+        property === "gridRow" ||
+        property === "gridColumn" ||
+        property === "gridRowStart" ||
+        property === "gridColumnStart" ||
+        property === "gridRowEnd" ||
+        property === "gridColumnEnd"
+      ) {
+        this.element.style[property] = `span ${value}${unit}`;
+      } else if (type === "grid-row" || type === "grid-col") {
+        this.element.style[property] = `repeat(${value}${unit}, 1fr))`;
+      } else if (type === "auto-grid-row" || type === "auto-grid-col") {
+        this.element.style[
+          property
+        ] = `repeat(auto-fit, minmax(${value}${unit}, 1fr))`;
       }
       // Default value and unit
       else {
@@ -79,7 +51,7 @@ makeTenoxUI.prototype.applyStyle = function (type, value, unit) {
 makeTenoxUI.prototype.applyStyles = function (className) {
   // Match all type and
   const match = className.match(
-    /(p|pt|pb|pr|pl|m|mt|mb|mr|ml|fs|fw|z|t|b|r|l|br|bw|bw-left|bw-right|bw-top|bw-bottom|w|w-mx|w-mn|h|h-mx|h-mn|fx|gap|ti|rt|backdrop-blur|ph|pv|mv|mh|text-space|word-space|line-height|radius-tl|radius-tr|radius-bl|radius-br|radius-top|radius-bottom|radius-left|radius-right|gs|blur|opa|scale|x|y)-(-?(?:\d+(\.\d+)?)|(?:[a-zA-Z]+(?:-[a-zA-Z]+)*(?:-[a-zA-Z]+)*)|(?:\[[^\]]+\]))([a-zA-Z%]*)/
+    /([a-zA-Z]+(?:-[a-zA-Z]+)*)-(-?(?:\d+(\.\d+)?)|(?:[a-zA-Z]+(?:-[a-zA-Z]+)*(?:-[a-zA-Z]+)*)|(?:\[[^\]]+\]))([a-zA-Z%]*)/
   );
 
   if (match) {
@@ -96,113 +68,8 @@ makeTenoxUI.prototype.applyStyles = function (className) {
 
 // TenoxUI Selector
 const AllClasses = document.querySelectorAll(
-  '[class*="p-"], [class*="pt-"], [class*="pb-"], [class*="pr-"], [class*="pl-"], [class*="m-"], [class*="mt-"], [class*="mb-"], [class*="mr-"], [class*="ml-"], [class*="fs-"], [class*="fw-"], [class*="z-"], [class*="t-"], [class*="b-"], [class*="r-"], [class*="l-"], [class*="br-"], [class*="bw-"], [class*="w-"], [class*="w-mx-"], [class*="w-mn-"], [class*="h-"], [class*="h-mx-"], [class*="h-mn-"], [class*="fx-"], [class*="gap-"], [class*="ti-"], [class*="rt-"], [class*="backdrop-blur-"], [class*="ph-"], [class*="pv-"], [class*="mv-"], [class*="mh-"], [class*="text-space-"], [class*="word-space-"], [class*="line-height-"], [class*="radius-tl-"], [class*="radius-tr-"], [class*="radius-bl-"], [class*="radius-br-"], [class*="radius-top-"], [class*="radius-bottom-"], [class*="radius-left-"], [class*="radius-right-"], [class*="radius-top-"], [class*="radius-bottom-"], [class*="radius-left-"], [class*="radius-right-"], [class*="gs-"], [class*="blur-"], [class*="opa-"], [class*="x-"], [class*="bw-left-"], [class*="bw-right-"], [class*="bw-top-"], [class*="bw-bottom-"], [class*="scale-"]'
+  '[class*="p-"], [class*="pt-"], [class*="pb-"], [class*="pr-"], [class*="pl-"], [class*="m-"], [class*="mt-"], [class*="mb-"], [class*="mr-"], [class*="ml-"], [class*="fs-"], [class*="fw-"], [class*="z-"], [class*="t-"], [class*="b-"], [class*="r-"], [class*="l-"], [class*="br-"], [class*="bw-"], [class*="w-"], [class*="w-mx-"], [class*="w-mn-"], [class*="h-"], [class*="h-mx-"], [class*="h-mn-"], [class*="fx-"], [class*="gap-"], [class*="ti-"], [class*="rt-"], [class*="backdrop-blur-"], [class*="ph-"], [class*="pv-"], [class*="mv-"], [class*="mh-"], [class*="text-space-"], [class*="word-space-"], [class*="line-height-"], [class*="radius-tl-"], [class*="radius-tr-"], [class*="radius-bl-"], [class*="radius-br-"], [class*="radius-top-"], [class*="radius-bottom-"], [class*="radius-left-"], [class*="radius-right-"], [class*="radius-top-"], [class*="radius-bottom-"], [class*="radius-left-"], [class*="radius-right-"], [class*="gs-"], [class*="blur-"], [class*="opa-"], [class*="x-"], [class*="bw-left-"], [class*="bw-right-"], [class*="bw-top-"], [class*="bw-bottom-"], [class*="scale-"], [class*="grid-row-"], [class*="grid-col-"], [class*="grid-item-row-"], [class*="grid-item-col-"]'
 );
-
-function TenoxColor() {
-  const setColor = (element, pattern, property, format) => {
-    const match = element.className.match(pattern);
-    if (match) {
-      element.style[property] = format(match);
-    }
-  };
-  // Class
-  const ColorableClass = document.querySelectorAll(
-    '[class*="bg-"], [class*="tc-"], [class*="border-"]'
-  );
-  ColorableClass.forEach((element) => {
-    // Recognize CSS Variable names. Variables name is defined first because in the last part is extended css color recognize.
-    setColor(
-      element,
-      /bg-([a-zA-Z0-9-]+)/,
-      "background",
-      (match) => `var(--${match[1]})`
-    );
-    setColor(
-      element,
-      /tc-([a-zA-Z0-9-]+)/,
-      "color",
-      (match) => `var(--${match[1]})`
-    );
-    setColor(
-      element,
-      /border-([a-zA-Z0-9-]+)/,
-      "borderColor",
-      (match) => `var(--${match[1]})`
-    );
-    // For RGB colors
-    setColor(
-      element,
-      /bg-rgb\((\d+),(\d+),(\d+)\)/,
-      "background",
-      (match) => `rgb(${match.slice(1, 4).join(",")})`
-    );
-    setColor(
-      element,
-      /tc-rgb\((\d+),(\d+),(\d+)\)/,
-      "color",
-      (match) => `rgb(${match.slice(1, 4).join(",")})`
-    );
-    setColor(
-      element,
-      /border-rgb\((\d+),(\d+),(\d+)\)/,
-      "borderColor",
-      (match) => `rgb(${match.slice(1, 4).join(",")})`
-    );
-    // For RGBA colors
-    setColor(
-      element,
-      /bg-rgba\((\d+),(\d+),(\d+),([01](\.\d+)?)\)/,
-      "background",
-      (match) => `rgba(${match.slice(1, 5).join(",")})`
-    );
-    setColor(
-      element,
-      /tc-rgba\((\d+),(\d+),(\d+),([01](\.\d+)?)\)/,
-      "color",
-      (match) => `rgba(${match.slice(1, 5).join(",")})`
-    );
-    setColor(
-      element,
-      /border-rgba\((\d+),(\d+),(\d+),([01](\.\d+)?)\)/,
-      "borderColor",
-      (match) => `rgba(${match.slice(1, 5).join(",")})`
-    );
-    // For HEX colors
-    setColor(
-      element,
-      /bg-([0-9a-fA-F]{3,6})/,
-      "background",
-      (match) => `#${match[1]}`
-    );
-    setColor(
-      element,
-      /tc-([0-9a-fA-F]{3,6})/,
-      "color",
-      (match) => `#${match[1]}`
-    );
-    setColor(
-      element,
-      /border-([0-9a-fA-F]{3,6})/,
-      "borderColor",
-      (match) => `#${match[1]}`
-    );
-    // All Posible Color In Css
-    setColor(
-      element,
-      /bg-([a-zA-Z0-9-]+)/,
-      "background",
-      (match) => `${match[1]}`
-    );
-    setColor(element, /tc-([a-zA-Z0-9-]+)/, "color", (match) => `${match[1]}`);
-    setColor(
-      element,
-      /border-([a-zA-Z0-9-]+)/,
-      "borderColor",
-      (match) => `${match[1]}`
-    );
-  });
-}
 
 AllClasses.forEach((element) => {
   const classes = element.classList;
@@ -211,5 +78,3 @@ AllClasses.forEach((element) => {
     makeTx.applyStyles(className);
   });
 });
-
-TenoxColor();